部分修正 - 将 brotli 库直接加入项目，因为低版本 Shadow 或者 jarJar 这样的工具都不起作用，而且 brotli 代码量小，且是 MIT 协议 - 修正 ID 读取问题，应当是 long 形式读取

+++ src/main/java/com/github/tartaricacid/bakadanmaku/site/bilibili/RoomInfo.java
@@ -49,6 +49,6 @@ public class RoomInfo {
         if (extractRoomId == null) {
             return null;
         }
-        return new RoomInfo(Integer.parseInt(extractRoomId), Integer.parseInt(ownerId == null ? "0" : ownerId));
+        return new RoomInfo(Long.parseLong(extractRoomId), Long.parseLong(ownerId == null ? "0" : ownerId));
     }
 }
